requirements.txt
numpy>=1.14.5
tensorboardX>=1.2
pytorch==0.4.0
tifffile==0.14.0
h5py>=2.7.1
scipy>=1.1.0
pybind11>=2.2.3
psutil>=5.4.7
jupyter==1.0.0
matplotlib==2.2.3
pillow==6.0.0

